done. seems to work probably
All done. Seems to work reasonably, but hard to tell with this kind of thing. It gives high chances for bad risk factors, 0 for good, and low but positive for less bad. The splits it makes also seem reasonable, for example the first one is splitting people with low age off into a low risk factor group. Who knows, could use more testing.